Moved Edge (->FmiEdge) and Node (->FmiNode) from subpackage model.fmi to model
